MAINT: Drop internal-move shims, export ToolCall, merge Conversation
Address PR review: the deep-path back-compat shims for purely internal
pyrit.models relocations add confusion rather than value, since every symbol
stays re-exported from pyrit.models. Delete the lingering silent shims at
attack_result.py, message.py, message_piece.py, and strategy_result.py so the
single canonical import (from pyrit.models import X) is the only supported form.
Warning shims remain only where a symbol genuinely left the package (the
pyrit.memory.storage moves).

Export ToolCall from pyrit.models so it is reachable via the package root like
every other messages symbol, and fold the one-class conversation.py module into
conversations.py to remove the near-duplicate module name.
