+++ tests/unit/models/test_chat_message.py
@@ -6,8 +6,7 @@ import json
 import pytest
 from pydantic import ValidationError
 
-from pyrit.models import ChatMessage, ChatMessagesDataset
-from pyrit.models.messages.chat_message import ToolCall
+from pyrit.models import ChatMessage, ChatMessagesDataset, ToolCall
 
 
 def test_tool_call_init():
