+++ tests/unit/models/test_strategy_result.py
@@ -1,11 +1,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-import warnings
-
 import pytest
 
-from pyrit.models.results.strategy_result import StrategyResult
+from pyrit.models import StrategyResult
 
 
 class ConcreteResult(StrategyResult):
@@ -39,14 +37,3 @@ def test_strategy_result_duplicate_preserves_type():
 def test_strategy_result_forbids_extra_fields():
     with pytest.raises(ValueError):
         ConcreteResult(value="hello", count=1, unexpected="boom")
-
-
-def test_strategy_result_shim_reexports_same_class_silently():
-    """The old import path must re-export the identical class without warning."""
-    with warnings.catch_warnings(record=True) as caught:
-        warnings.simplefilter("always")
-        from pyrit.models.strategy_result import StrategyResult as ShimStrategyResult
-
-    assert ShimStrategyResult is StrategyResult
-    deprecation_warnings = [w for w in caught if issubclass(w.category, DeprecationWarning)]
-    assert len(deprecation_warnings) == 0, "Shim import must be silent"
